move fetch to dashbaord component
