+++ wireframe-frontend/src/containers/dashboard.js
@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from 'react';
+import React, { Component } from 'react';
 import { BrowserRouter as Router, Route, Link, Switch } from "react-router-dom";
 import '../App.css';
 import { Button } from 'reactstrap';
@@ -7,12 +7,23 @@ import fetch from 'isomorphic-fetch'
 import runtimeEnv from '@mars/heroku-js-runtime-env'
 
 export default class Dashboard extends Component {
-  constructor(props){
+  constructor() {
     super()
+    this.state = {
+      data: ""
+    }
   }
+
+  componentDidMount() {
+    const url = runtimeEnv().REACT_APP_API_URL
+    fetch(url)
+      .then( res => res.json() )
+      .then( json => this.setState({ data: json }) )
+  }
+
   render(){
     return (
-      <Button>Dashboard</Button>
+      <Button>Dashboard: {this.state.data}</Button>
     )
   }
 }
