+++ wireframe-frontend/src/containers/login.js
@@ -1,4 +1,4 @@
-import React, { Component, Fragment } from 'react';
+import React, { Component } from 'react';
 import { BrowserRouter as Router, Route, Link, Switch } from "react-router-dom";
 import '../App.css';
 import { Button } from 'reactstrap';
